expectation_management_tool/requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.18.18
botocore==1.21.18
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
daphne==3.0.2
Django==3.2.6
django-cors-headers==3.7.0
django-fcm==0.1.1
django-phonenumber-field==5.2.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.3
incremental==21.3.0
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.3
micawber==0.5.4
msgpack==1.0.2
phonenumbers==8.12.28
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pyunfurl==0.1.5
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritools==3.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
zope.interface==5.4.0